refactor: next.config.js base on the project root
